Update Version.h
Fix invalid characters for encoding. Resolves the following warning in Arduino IDE:
"Version.h" contains unrecognised characters. If this code was created with an older version of Arduino, you may need to use Tools -> Fix Encoding & Reload to update the sketch to use UTF-8 encoding. If not, you may need to delete the bad characters to get rid of this warning.